requirements.txt
tensorflow>=2.8.0
numpy>=1.19.2
pandas>=1.2.4
Pillow>=8.2.0
matplotlib>=3.4.2
scikit-learn>=0.24.2 